Add the new Unsafe primitives to update values in a private buffer
and test method to test if a field is flattened and if an array class
is a flattened array.

The Unsafe APIs for values include the following:

boolean Unsafe::isFlattened(Field f);
boolean Unsafe::isFlattenedArray(Class<?> arrayClass);
<V> long valueHeaderSize(Class<V> vc);
<V> V makePrivateBuffer(V value);
<V> V finishPrivateBuffer(V value);
<V> V getValue(Object o, long offset, Class<?> vc);
<V> void putValue(Object o, long offset, Class<?> vc, V v);

makePrivateBuffer creates a value instance whose layout and
contents is exactly the same as the input value and marks it in
larval state and finishPrivateBuffer will exit the larval state.
In between Unsafe.putXXX on the buffer will update the larval
value instance.