Class C attempts to access a protected member D::m inherited from
[nestmates] Lambda proxy uses live MethodHandle for a protected member inherited from its superclass in a different package.
its supertype in a different package. The lambda proxy class HC is not a subtype
of D and D::m is not inaccessible to HC.  C may not necessarily have the bridge
for example D::m was public member when C was compiled and later D::m is changed
from public to protected in a separation compilation.  D::m is accessible to C
with a receiver object of C.  Changing D::m to protected is a supported binary
compatibility.  JDK-8199386 should be explored as a long-term solution that
a framework can use Lookup::in to teleport from HC to C to obtain a method
handle for D::m.