8264414: [lworld] [AArch64] TestBufferTearing.java fails with C1
We see failures like this on AArch64 when MyValue.incrementAndCheck() is
compiled with C1:

  java.lang.RuntimeException: Inconsistent field values: expected 0 to equal 675128
        at jdk.test.lib.Asserts.fail(Asserts.java:594)
        at jdk.test.lib.Asserts.assertEquals(Asserts.java:205)
        at jdk.test.lib.Asserts.assertEQ(Asserts.java:178)
        at compiler.valhalla.inlinetypes.MyValue.incrementAndCheck(TestBufferTearing.java:81)
        at compiler.valhalla.inlinetypes.TestBufferTearing$Runner.run(TestBufferTearing.java:124)

The barrier that is usually inserted on return from a method that wrote
final fields should be sufficient to prevent another thread seeing the
zero-initialised intermediate state.  However this barrier isn't
inserted at the moment because method()->is_object_constructor() is
false for primitive class constructors.

C2 has a similar guard around the memory barrier in Parse::do_exits().
I'm not sure if that needs amending as well but I've not seen any
failures due to it.