FP16 isNaN, isFinite, isInfinite intrinsics without Reinterpret nodes
This patch removes the ReinterpretS2HF nodes in the mid-end during the
generation of isNaNHF,isFiniteHF and isInfiniteHF nodes.

Performance results for this patch on an aarch64 machine -

Benchmark               Gain over baseline      Gain over default
FP16Ops.isFiniteHF      1.29                    1.85
FP16Ops.isInfiniteHF    1.28                    1.90
FP16Ops.isNaNHF	65504   1.45                    1.58

The baseline patch generates floating point FP16 instructions and the
default is where no FP16 intrinsics are used and FP32 instructions are
generated.

Gain : thrpt of this patch / thrpt of either baseline or default

Tested FP16Ops.isInfiniteHF test on x86 and the performance is 2.6x
better over the default case (which converts FP16 to FP32 and uses the
vfpclass instruction).

The JMH tests are added in this patch.